requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
psycopg2==2.8.4
PyJWT==1.7.1
SQLAlchemy==1.3.13
Werkzeug==0.16.1
